Initialize request body from req.blob()
There seems to be a lot of bugs in bun regarding streaming request
bodies, which matches what we have seen in this application
(sometimes apparent missing request bodies even with correct
content-length). Bun.js is working on implementing request.clone but in
the mean time the best is to parse the request body as a blob and pass
the whole body in memory to the new request.